zoom.sh: Lie to Zoom about the desktop in Wayland
Zoom employs a check for Wayland screensharing to only allow it when
it detects that GNOME is the running desktop. However, since Zoom
5.11.0, Wayland screensharing is done through the XDG desktop portal
mechanism, and thus is desktop independent.

Unfortunately, the check is still in place, and in order to allow
Wayland screen sharing to work on KDE Plasma and other Wayland
environments, we need to lie to Zoom to allow it through.

Thus, we set XDG_CURRENT_DESKTOP to "gnome" whenever we detect we
are running Zoom in a Wayland environment to satisfy the check.

Also, add the wayland privilege to the app, since Zoom's Qt supports
it now.